+++ packages/model/system_restore_access.go
@@ -5,7 +5,7 @@ type SystemRestoreAccess struct {
 	CitizenID int64  `gorm:"not null"`
 	StateID   int64  `gorm:"not null"`
 	Active    int64  `gorm:"not null"`
-	Time      int32  `gorm:"not null"`
+	Time      int64  `gorm:"not null"`
 	Close     int64  `gorm:"not null"`
 	Secret    string `gorm:"not null"`
 	RbID      int64  `gorm:"not null"`
